\name{createbrickWS}

\docType{methods}

\alias{createbrickWS}

\title{Create a brick (raster object) from the downloaded MODIS data}

\description{
Get the available Dates of images at the a location for a specific MODIS product available via the MODIS webservice tool
}

\usage{createbrickWS(result)}

\arguments{
	\item{result}{ data from MODIS webservice }
}

\value{
	Extracting and processing MODIS data into a raster brick object.
}

\details{
Please see \url{http://daac.ornl.gov/MODIS/MODIS-menu/modis_webservice.html} for more details about the service.
}

\author{
  Jan Verbesselt
}

\examples{
## this is a helper function
}
