% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarthdataLogin.R
\name{EarthdataLogin}
\alias{EarthdataLogin}
\title{Create File with Earthdata Login Credentials}
\usage{
EarthdataLogin(service = "Earthdata", usr = NULL, pwd = NULL)
}
\arguments{
\item{service}{\code{character}. Search service for MODIS data, defaults to 
\code{"Earthdata"} which is currently the only option available.}

\item{usr, pwd}{Login credentials as \code{character}. If \code{NULL}, 
username and password are read from the terminal.}
}
\value{
An \code{invisible()} absolute file path of the created \code{.netrc} file as 
\code{character}.
}
\description{
Create a hidden \code{.netrc} file with Earthdata login credentials in your 
home directory. The information included therein is used to login to 
\url{urs.earthdata.nasa.gov} which is a mandatory requirement in order to 
download MODIS data from LPDAAC, LAADS and NSIDC (see also 
\code{\link{MODISoptions}}).
}
\examples{
\dontrun{
EarthdataLogin()
}

}
\seealso{
\itemize{
\item{\url{http://docs.opendap.org/index.php/DAP_Clients_-_Authentication#LDAP} 
(section 2.2)}
\item{\url{https://github.com/MatMatt/MODIS/issues/10}}
\item{\url{https://wiki.earthdata.nasa.gov/display/EL/How+To+Access+Data+With+cURL+And+Wget}}
}
}
\author{
Matteo Mattiuzzi and Florian Detsch
}
