% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSize.R
\name{fileSize}
\alias{fileSize}
\title{Get Size of File(s)}
\usage{
fileSize(file, units = c("B", "KB", "MB", "GB", "TB"))
}
\arguments{
\item{file}{\code{character} vector of file(s) with path.}

\item{units}{\code{character}, defaults to \code{"B"}. Currently available options
are \code{c("B", "KB", "MB", "GB", "TB")} for bites, kilo-, mega-, giga- and
terabytes.}
}
\value{
A \code{numeric} vector of the same length as 'file' (in 'units'). Note that
directories are excluded.
}
\description{
Get the size of any file.
}
\examples{
\dontrun{
fileSize(list.files("./"))
}

}
\author{
Matteo Mattiuzzi
}
