% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGranule.R
\name{getGranule}
\alias{getGranule}
\title{Get MODIS Swath Granules}
\usage{
getGranule(product, collection = NULL, begin = NULL, end = NULL,
  DayNightFlag = c("D", "N", "B", "X"), tileH = NULL, tileV = NULL,
  extent = NULL, ...)
}
\arguments{
\item{product, collection, begin, end, tileH, tileV, extent, ...}{See 
\code{\link[MODIS]{getHdf}}.}

\item{DayNightFlag}{A \code{character} vector of allowed day/night flags.
This can be an arbitrary combination of \code{"D"} (day), \code{"N"} (night), 
\code{"B"} (both), \code{"X"} (not designated). By default, all flags are 
accepted.}
}
\value{
Identified granules as \code{character}.
}
\description{
Get MODIS swath granules for a specific geographic area, time period and 
(optionally) time of day.
}
\examples{
\dontrun{
data(meuse)
pts = sf::st_as_sf(meuse, coords = c("x", "y"), crs = 28992)

grn = getGranule("MOD14", begin = "2000360", end = "2001003"
                 , extent = pts, DayNightFlag = "D")
}

}
\seealso{
\code{\link[MODIS]{getHdf}}, \code{\link[MODIS]{getTile}}.
}
\author{
Florian Detsch
}
