\name{getProductsWS}

\docType{methods}

\alias{getProductsWS}

\title{Check the avalability of the MODIS products via the MODIS webservice tool}

\description{
	Check the avalability of the MODIS products via the MODIS webservice tool.
}

\usage{getProductsWS()}

\value{
	A list with the available MODIS products
}

\author{
  Jan Verbesselt
}

\examples{
\dontrun{
	#	Please install the required packages
	getProductsWS()
}
}
