% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdaacLogin.R
\name{lpdaacLogin-deprecated}
\alias{lpdaacLogin-deprecated}
\title{Create File with Earthdata Login Credentials}
\usage{
lpdaacLogin(server = "LPDAAC")
}
\arguments{
\item{server}{\code{character}. MODIS file server, defaults to \code{"LPDAAC"} which is
currently the only option available.}
}
\value{
Invisible.
}
\description{
Create a hidden .netrc file with Earthdata Login credentials in your home
directory. If your priority server for MODIS file download is LP DAAC (see
also \code{\link[=MODISoptions]{MODISoptions()}}), these are subsequently used to automatically login to
\url{https://urs.earthdata.nasa.gov/} and download required files.
}
\examples{
\dontrun{
lpdaacLogin()
}

}
\seealso{
\itemize{
\item \url{https://docs.opendap.org/index.php/DAP_Clients_-_Authentication#LDAP}
(section 2.2)
\item \url{https://github.com/fdetsch/MODIS/issues/10}
}

\link{MODIS-deprecated}
}
\author{
Matteo Mattiuzzi and Florian Detsch
}
\keyword{internal}
