% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdaacLogin.R
\name{lpdaacLogin}
\alias{lpdaacLogin}
\title{Create File with Earthdata Login Credentials}
\usage{
lpdaacLogin(server = "LPDAAC")
}
\arguments{
\item{server}{\code{character}. MODIS file server, defaults to 
\code{"LPDAAC"} which is currently the only option available.}
}
\value{
\code{invisible()}.
}
\description{
Create a hidden .netrc file with Earthdata login credentials in your home 
directory. If your priority server for MODIS file download is LPDAAC (see 
also \code{\link{MODISoptions}}), these are subsequently used to 
automatically login to \url{urs.earthdata.nasa.gov} and download required 
files.
}
\examples{
\dontrun{
lpdaacLogin()
}

}
\author{
Matteo Mattiuzzi
}
\seealso{
\itemize{
\item{\url{http://docs.opendap.org/index.php/DAP_Clients_-_Authentication#LDAP} 
(section 2.2)}
\item{\url{https://github.com/MatMatt/MODIS/issues/10}}
}
}

