% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODISswath.R
\name{makeFootprintSwath}
\alias{makeFootprintSwath}
\alias{listGeometa}
\title{Get MODIS Swath Granules}
\usage{
listGeometa(platform = c("terra", "aqua"), begin = NULL,
  end = Sys.Date(), collection = 6, forceRefresh = FALSE, ...)
}
\arguments{
\item{platform}{\code{Character} of length 1, either MODIS "terra" or "aqua"}

\item{collection}{integer or character, valid values are collection 5 or 6}

\item{...}{currently only works for the MODIS outDirPath root directory.}

\item{DayNightFlag}{A \code{character} vector of allowed day/night flags.
This can be an arbitrary combination of \code{"D"} (day), \code{"N"} (night), 
\code{"B"} (both), \code{"X"} (not designated). By default, all flags are 
accepted.}

\item{removeDatecrossers}{Default TRUE, remove granules that cross the 
dateline as current solution warps polygons accross the globe (any solutions
on how to solve this issue are welcome!)}
}
\value{
Identified granules as \code{character}.
}
\description{
Get MODIS swath granules for a specific geographic area, time period and 
(optionally) time of day.
}
\details{
The retriveval of Spath data follows a three step approach: 
1) download and store CSV summary files. These yearly summary files contain 
the filenames of each geometa: MYD03 or MOD03file.
2) base in hte CSV summary file download and store selected geometa files
3) Extract Swath footprint geometries from Geometafiles and intersect it 
with the extent.
As this function uses caching once files are downloaded if becomes prety fast.
}
\examples{
\dontrun{

data(meuse)
pts = sf::st_as_sf(meuse, coords = c("x", "y"), crs = 28992)

begin <- '2017001'
end   <- '2017010'
 
fp <- makeFootprintSwath(extent=pts, platform='terra', 
            begin = begin, end = end, DayNightFlag = "D")

plot(fp$geometry)
fp

}

}
\author{
Florain Detsch and Matteo Mattiuzzi
}
