% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS-package.R
\docType{package}
\name{MODIS-package}
\alias{MODIS}
\alias{MODIS-package}
\title{MODIS Acquisition and Processing}
\description{
MODIS Acquisition and Processing
}
\details{
Download and processing functionality for the Moderate Resolution Imaging
Spectroradiometer (MODIS). The package provides automated access to the
global online data archives (LP DAAC and LAADS) and processing capabilities
such as file conversion, mosaicking, subsetting and time series filtering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdetsch/MODIS}
  \item Report bugs at \url{https://github.com/fdetsch/MODIS/issues}
}

}
\author{
Matteo Mattiuzzi, Florian Detsch
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
