\name{MODIS_Products}
\alias{MODIS_Products}
\docType{data}
\title{
MODLAND product table 
}
\description{
MODLAND product description table from: \url{https://lpdaac.usgs.gov/lpdaac/products/modis_products_table}
}
\usage{data(MODIS_Products)}

\source{\url{https://lpdaac.usgs.gov/lpdaac/products/modis_products_table}}

\examples{
data(MODIS_Products)
MODIS_Products
}
\keyword{datasets}
