% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{MODISextent-class}
\alias{MODISextent-class}
\title{Class MODISextent}
\description{
An object of class \code{MODISextent}, typically created through 
\code{\link{getTile}}.
}
\section{Slots}{

\describe{
\item{\code{tile}}{MODIS tile ID as \code{character}.}

\item{\code{tileH}}{MODIS horizontal tile ID as \code{integer}.}

\item{\code{tileV}}{MODIS vertical tile ID as \code{integer}.}

\item{\code{extent}}{\code{Extent} information, see \code{\link{getTile}}.}

\item{\code{system}}{Sensor system as \code{character}.}

\item{\code{target}}{If applicable, a \code{list} with additional target information.}
}}

