% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{MODISproduct-class}
\alias{MODISproduct-class}
\title{Class MODISproduct}
\description{
An object of class \code{MODISproduct}, typically created as the 
\code{\link{invisible}} output of \code{\link[MODIS]{getProduct}}.
}
\section{Slots}{

\describe{
\item{\code{request}}{User request as \code{character}.}

\item{\code{PF1,PF2,PF3,PF4}}{Platform specific path feature for LP DAAC, LAADS, NTSG 
and NSIDC as \code{character}.}

\item{\code{PD}}{Product specific code number as 2-digit \code{character}.}

\item{\code{PLATFORM}}{Satellite platform on which MODIS sensor is mounted; one of 
\code{c("Terra", "Aqua")}.}

\item{\code{TYPE}}{Product type; one of \code{c("Tile", "CMG", "Swath")}.}

\item{\code{PRODUCT}}{MODIS product identified from 'request' as \code{character}.}

\item{\code{SOURCE}}{Product specific MODIS download server(s) as \code{list}.}

\item{\code{CCC}}{Product specific MODIS data collection as 3-digit \code{character}.}
}}

