% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preStack.R
\name{preStack}
\alias{preStack}
\title{Organise (MODIS) Files in Preparation for Stacking}
\usage{
preStack(pattern = "*", path = "./", files = NULL, timeInfo = NULL)
}
\arguments{
\item{pattern}{Regular expression passed to \code{\link{list.files}}}

\item{path}{\code{character}. Location of MODIS files to stack.}

\item{files}{\code{character} vector of filenames. If provided, arguments 
\code{pattern} and \code{path} are ignored.}

\item{timeInfo}{Ouput from \code{\link{orgTime}}.}
}
\value{
A \code{character} vector of filenames within the query. If \code{timeInfo} 
is provided, filenames are sorted and subsetted by date.
}
\description{
This function lets you sort a vector of filenames according to date. It is 
thought to be used on results from \code{\link{runGdal}} or 
\code{\link{runMrt}}.
}
\examples{
# see Examples in ?smooth.spline.raster

}
\author{
Matteo Mattiuzzi
}

