% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSize.R
\name{fileSize}
\alias{fileSize}
\title{Get Size of File(s)}
\usage{
fileSize(file, units = "B")
}
\arguments{
\item{file}{\code{character} vector of file(s) with path.}

\item{units}{\code{character}, defaults to "B". Currently available options 
are "B", "KB", "MB", "GB" or "TB" for bites, kilo-, mega-, giga- and terabytes.}
}
\value{
\code{numeric} vector of the same length as \code{file} (in \code{units}). 
Note that directories are excluded.
}
\description{
Function for getting size of any files.
}
\examples{
\dontrun{
fileSize(list.files("./"))
}

}
\author{
Matteo Mattiuzzi
}
