% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHdf.R
\docType{methods}
\name{getHdf}
\alias{getHdf}
\alias{getHdf,character-method}
\alias{getHdf,list-method}
\alias{getHdf,missing-method}
\title{Create or update a local subset of global online MODIS grid data pool}
\usage{
\S4method{getHdf}{missing}(product, begin = NULL, end = NULL,
  tileH = NULL, tileV = NULL, extent = NULL, collection = NULL,
  wait = 0.5, checkIntegrity = TRUE, forceDownload = TRUE, ...)

\S4method{getHdf}{character}(HdfName, wait = 0.5, checkIntegrity = TRUE,
  ...)

\S4method{getHdf}{list}(HdfName, wait = 0.5, checkIntegrity = TRUE, ...)
}
\arguments{
\item{product}{\code{character}. MODIS grid product to be downloaded, see 
\code{\link{getProduct}}. Use dot notation to address Terra and Aqua products 
(e.g. \code{M.D13Q1}).}

\item{begin}{\code{character}. Begin date of MODIS time series, see 
\code{\link{transDate}} for formatting.}

\item{end}{\code{character}. End date, compatible with future dates for 
continuous updates via scheduled jobs.}

\item{tileH}{\code{numeric} or \code{character}. Horizontal tile number(s), 
see \code{\link{getTile}}.}

\item{tileV}{\code{numeric} or \code{character}. Vertical tile number(s), 
see \code{tileH}.}

\item{extent}{See Details in \code{\link{getTile}}.}

\item{collection}{\code{character} or \code{integer}. Desired MODIS product 
collection, see MODIS pages or \code{\link{getCollection}} for more information.}

\item{wait}{\code{numeric}. Inserts a break (in seconds) after every internal 
call to \code{\link{download.file}} or \code{\link{getURL}}, which reduces 
the chance of FTP connection errors that frequently occur after many requests.}

\item{checkIntegrity}{\code{logical}. If \code{TRUE} (default), the size of 
each downloaded file is checked. In case of inconsistencies, the function 
tries to re-download broken files.}

\item{forceDownload}{\code{logical}. If \code{TRUE} (default), try to 
download data irrespective of whether online information could be retrieved 
via \code{MODIS:::getStruc} or not.}

\item{...}{Arguments found in \code{\link{MODISoptions}}, sections 'STORAGE' 
and 'DOWNLOAD'.}

\item{HdfName}{\code{character} vector or \code{list}. Full HDF file name(s) 
to download a small set of files. If specified, other file-related parameters 
(i.e., \code{begin}, \code{end}, \code{collection}, etc.) are ignored.}
}
\value{
An invisible vector of downloaded data and paths.
}
\description{
Create or update a local user-defined subset of the global MODIS grid data 
archive. Based on user-specific parameters the function checks in the local 
archive for available data and downloads missing data from the online MODIS 
data pool. When run in a schedule job, the function manage the continuous 
update of the local MODIS data archive.
}
\examples{
\dontrun{
# one or more specific file (no regular erpression allowed here)
a <- getHdf(c("MYD11A1.A2009001.h18v04.006.2015363221538.hdf", 
              "MYD11A1.A2009009.h18v04.006.2015364055036.hdf", 
              "MYD11A1.A2009017.h18v04.006.2015364115403.hdf"))

# DO NOT RUN THE NEXT EXAMPLE (unless you need this data)!!!
# Get all MODIS TERRA and AQUA 11A1 beginning from 18. October 2011 up to today.
# (Can be ran in a sceduled job, for daily archive update)
# getHdf(product="M.D11A1",begin="2011.10.18",tileH=18:19,tileV=4) 

# same Tiles with a LIST extent.
# approximatley 21 mB!
Austria <- list(xmax=17.47,xmin=9.2,ymin=46.12,ymax=49.3)
b <- getHdf(product="M.D11A1",begin="2009001",end="2009-01-02",extent=Austria)
b

# require(mapdata)
getHdf(product="M.D11A1",begin="2009.01.01",end="2009002",extent="austria")
# without specification of the extent... interactive selection see: "getTile()"
c <- getHdf(product="M.D11A1",begin="2009.01.01",end="2009002")
c

# SRTM data
# SRTM server limits downloads! (It seams to be blocked, but (normally) it continues after a while,
# you have to be patient!).
# The files are zips, this function only performs the download!
d <- getHdf(product="SRTM",extent="austria")  
d
}

}
\author{
Matteo Mattiuzzi
}
\references{
MODIS data is obtained through the online Data Pool at the NASA Land 
Processes Distributed Active Archive Center (LP DAAC), USGS/Earth Resources 
Observation and Science (EROS) Center, Sioux Falls, South Dakota 
\url{https://lpdaac.usgs.gov/get_data}.

SRTM data is obtained through CGIAR-CSI and mirror servers. For the use this 
data please read \url{http://srtm.csi.cgiar.org/SELECTION/SRT_disclaimer.htm}.
Jarvis A., H.I. Reuter, A.  Nelson, E. Guevara, 2008, Hole-filled seamless 
SRTM data V4, International Centre for Tropical Agriculture (CIAT), available 
from \url{http://srtm.csi.cgiar.org}.
}

