% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS-deprecated.R, R/lpdaacLogin.R
\name{MODIS-deprecated}
\alias{MODIS-deprecated}
\alias{lpdaacLogin}
\title{Deprecated Functions in \pkg{MODIS}}
\usage{
lpdaacLogin(server = "LPDAAC")
}
\description{
The functions listed below are deprecated and will be defunct in the near
future. If applicable, alternative functions with similar or even identical
functionality are mentioned. Help pages for deprecated functions are
available at \code{help("MODIS-deprecated")}.
\itemize{
\item \code{\link[=lpdaacLogin-deprecated]{lpdaacLogin-deprecated()}}: use \link{EarthdataLogin} instead.
}
}
\section{\code{lpdaacLogin}}{

For \code{\link[=lpdaacLogin]{lpdaacLogin()}}, use \code{\link[=EarthdataLogin]{EarthdataLogin()}} instead.
}

\keyword{internal}
