% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS-package.R
\docType{package}
\name{MODIS-package}
\alias{MODIS-package}
\title{MODIS Acquisition and Processing}
\description{
MODIS Acquisition and Processing
}
\details{
Download and processing functionalities for the Moderate Resolution Imaging 
Spectroradiometer (MODIS). The package provides automated access to the 
global online data archives (LP DAAC and LAADS) and processing capabilities 
such as file conversion, mosaicking, subsetting and time series filtering.
}
\author{
Matteo Mattiuzzi, Jan Verbesselt, Tomislav Hengl, Anja Klisch, 
Forrest Stevens, Steven Mosher, Bradley Evans, Agustin Lobo, Florian Detsch 
\cr
\cr
\emph{Maintainer:} Florian Detsch \email{fdetsch@web.de}
}
\keyword{package}
