\name{MODIS_Products}
\alias{MODIS_Products}
\docType{data}
\title{
Internal Data description table 
}
\description{Defines data products usable in MODIS package.}

\usage{data(MODIS_Products)}

\source{
MODIS data information taken from: \url{https://lpdaac.usgs.gov/lpdaac/products/modis_products_table}\cr
MERIS data description taken from: \url{ftp://ionia2.esrin.esa.int/CULTURE-MERIS_PDM_2.0.pdf}
}

\examples{
data(MODIS_Products)
MODIS_Products
getProduct()

}
\keyword{datasets}
