% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preStack.R
\name{preStack}
\alias{preStack}
\title{Organize (MODIS) Files in Preparation for Stacking}
\usage{
preStack(pattern = "*", path = "./", files = NULL, timeInfo = NULL)
}
\arguments{
\item{pattern}{Regular expression passed to \code{\link[=list.files]{list.files()}}}

\item{path}{\code{character}. Location of MODIS files to stack.}

\item{files}{\code{character} vector of file names. If provided, arguments
'pattern' and 'path' are ignored.}

\item{timeInfo}{Output from \code{\link[=orgTime]{orgTime()}}.}
}
\value{
A \code{character} vector of file names within the query. If 'timeInfo' is
provided, file names are sorted and subsetted by date.
}
\description{
This function lets you sort a vector of file names according to date. It is
thought to be used on results from \code{\link[=runGdal]{runGdal()}} or \code{\link[=runMrt]{runMrt()}}.
}
\examples{
# see Examples in ?smooth.spline.raster

}
\author{
Matteo Mattiuzzi
}
